+++ src/entity/Bin.java
@@ -6,13 +6,14 @@ import java.util.List;
 public class Bin {
     private static List<Item> items = null;
     public static int COUNTER = 0;
-    public static int PRICE = 0;
+    public static Double PRICE = 0D;
 
 
     private Bin() {
     }
 
     public static List<Item> getItems() {
+
         if (items == null) {
             items = new ArrayList<>();
         }
