+++ src/entity/Item.java
@@ -1,10 +1,11 @@
 package entity;
 
+
 public class Item {
 
 
     private String name;
-    private int price;
+    private Double price;
     private long id;
 
     public Item() {
@@ -19,11 +20,11 @@ public class Item {
         this.name = name;
     }
 
-    public int getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
